+++ Examples/YandexMobileAdsExample/YandexMobileAdsExampleUITests/Pages/GDPRSettingsPage.swift
@@ -1,0 +1,13 @@
+import XCTest
+
+struct GDPRSettingsPage: PageObject {
+    var consentSwitch: XCUIElement { element(id: GDPRAccessibility.userConsentSwitch, type: .switch) }
+    var resetButton: XCUIElement { element(id: GDPRAccessibility.resetButton, type: .button) }
+    
+    func tapReset() {
+        step("Tap reset") {
+            resetButton.tap()
+        }
+    }
+}
+