+++ changelog/mediation/mobileads-mytarget/CHANGELOG.md
@@ -2,6 +2,14 @@
 
 All notable changes to this project will be documented in this file.
 
+## Version 5.20.1.0
+
+### Updated
+
+* Added support for MyTarget SDK version 5.20.1
+* Updated minimum supported MyTarget SDK version to 5.20.1
+* Updated minimum supported Yandex Mobile Ads SDK version to 6.4.0
+
 ## Version 5.20.0.0
 
 ### Updated